README.md
This is a general repository for the numerous ongoing projects that I currently have.
Please feel free to take a gander, but if you plan on utilizing any of it or are inspired by it, please reach out and let me know. I'm always happy to hear how others are innovating. Big believer in arguing over points that result in a better end-product.